requirements.txt
cycler==0.11.0
flake8==4.0.1
fonttools==4.28.5
kiwisolver==1.3.2
matplotlib==3.5.1
mccabe==0.6.1
numpy==1.22.1
packaging==21.3
pandas==1.3.5
Pillow==9.0.0
pkg_resources==0.0.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.6
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.10.0
six==1.16.0
